// src/Models/Seo.ts
export interface ISeo {
  metaTitle: string;
  metaKeyword: string;
  metaDescription: string;
  favicon: string;
  featuredImage: string;
  metaTags: string;
  robotTxt: string;
  excludeSite: string;
  sitemapHtml: string;
  sitemapXml: string;
  sitemapPagesXml: string;
  nftCollectionXml: string;
  nftItemsXml: string;
  usersXml: string;
  collectionRSSTitle: string;
  collectionRSSUrl: string;
  collectionRSSDescription: string;
  showCollectionFeed: string;
  assetsRSSTitle: string;
  assetsRSSUrl: string;
  assetsRSSDescription: string;
  showAssetsFeed: string;
  showAssetsSchema: string;
  showCollectionSchema: string;
}